listener/pyproject.toml
[tool.poetry]
name = "listener"
version = "0.1.0"
description = "Listener script that calls xrandr when accelerometer data arrives"
authors = ["sireliah"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyserial = "^3.5"

[tool.poetry.group.dev.dependencies]
notebook = "^7.0.6"
matplotlib = "^3.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
